move data transformation methods to separate file
